重构部分TraverseTool，结构更清晰，testcase01通过

+++ src/traverse_tool.py
@@ -1,40 +1,41 @@
 from abc import ABC, abstractmethod
 from pathlib import Path
-
+from tkinter import W
+from common_tool import ignore_list
 
 class TraverseTool():
     def traverse_path(self, parent: Path):
-        for child in parent.iterdir():
-            if self.check_ignore(child):
-                if child.is_dir():
+        if parent.is_file():
+            self.handle_file(parent)
+        else:
+            for child in parent.iterdir():
+                if self.check_ignore(child):
+                    if child.is_dir():
 
-                    # Check empty
-                    flag = False
-                    for _ in child.iterdir():
-                        flag = True
-                        break
+                        # Check empty
+                        flag = False
+                        for _ in child.iterdir():
+                            flag = True
+                            break
 
-                    if flag: # not empty
-                        self.check_dir(child)
+                        if flag: # not empty
+                            self.check_dir(child)
+                        else:
+                            child.rmdir()
+                    elif child.is_file():
+                        self.handle_file(child)
                     else:
-                        child.rmdir()
-                elif child.is_file():
-                    # print('FILE ' + str(child), sep='\n')
-                    if child.suffix == '.zip':
-                        self.handle_zip(child)
-                    elif child.suffix == '.rar':
-                        self.handle_rar(child)
-                    else:
-                        self.handle_other(child)
-                else:
-                    pass
-                    # print('others')
-
+                        pass
     def check_ignore(self, path: Path) -> bool:
+        if path.name in ignore_list:
+            return False
         return True
 
     def check_dir(self, path: Path):
         self.traverse_path(path)
+
+    def handle_file(self, path: Path) -> None:
+        pass
 
     def handle_zip(self, file: Path):
         pass
